+++ mcp-server/jest.config.ts
@@ -1,5 +1,4 @@
-/** @type {import('ts-jest').JestConfigWithTsJest} */
-const config = {
+const config: import("ts-jest").JestConfigWithTsJest = {
   preset: "ts-jest/presets/js-with-ts-esm",
   testEnvironment: "node",
   extensionsToTreatAsEsm: [".ts"],
@@ -8,6 +7,7 @@ const config = {
     "\\.md$": "<rootDir>/src/__mocks__/markdownMock.js",
     "^@/clients/(.*)\\.js$": "<rootDir>/src/clients/$1",
     "^@/helpers/(.*)\\.js$": "<rootDir>/src/helpers/$1",
+    "@/test-helpers/(.*)\\.js$": "<rootDir>/src/test-helpers/$1",
     "^@/constants/(.*)\\.js$": "<rootDir>/src/constants/$1",
     "^@/umb-management-api/(.*)\\.js$":
       "<rootDir>/src/umb-management-api/api/$1",
