+++ mcp-server/src/test-helpers/create-snapshot-result.ts
