+++ mcp-server/src/umb-management-api/tools/dictionary/__tests__/get-dictionary-item.test.ts
@@ -2,7 +2,7 @@ import { getDictionaryByIdParams } from "@/umb-management-api/umbracoManagementA
 import GetDictionaryItemTool from "../get/get-dictionary-item.js";
 import { DictionaryBuilder } from "./helpers/dictionary-builder.js";
 import { DEFAULT_ISO_CODE } from "./helpers/dictionary-helper.js";
-import { createSnapshotResult } from "@/helpers/test-utils.js";
+import { createSnapshotResult } from "@/test-helpers/create-snapshot-result.js";
 import { jest } from "@jest/globals";
 import { BLANK_UUID } from "@/constants/constants.js";
