+++ mcp-server/src/umb-management-api/tools/document-type/__tests__/get-document-type-allowed-children.test.ts
@@ -1,7 +1,7 @@
 import GetDocumentTypeAllowedChildrenTool from "../get/get-document-type-allowed-children.js";
 import { DocumentTypeBuilder } from "./helpers/document-type-builder.js";
 import { DocumentTypeTestHelper } from "./helpers/document-type-test-helper.js";
-import { createSnapshotResult } from "@/helpers/test-utils.js";
+import { createSnapshotResult } from "@/test-helpers/create-snapshot-result.js";
 import { jest } from "@jest/globals";
 import { BLANK_UUID } from "@/constants/constants.js";
