+++ mcp-server/src/umb-management-api/tools/media-type/__tests__/get-media-type-by-ids.test.ts
@@ -1,6 +1,6 @@
 import { MediaTypeTestHelper } from "./helpers/media-type-helper.js";
 import GetMediaTypeByIdsTool from "../get/get-media-type-by-ids.js";
-import { createSnapshotResult } from "@/helpers/test-utils.js";
+import { createSnapshotResult } from "@/test-helpers/create-snapshot-result.js";
 import { jest } from "@jest/globals";
 import { MediaTypeBuilder } from "./helpers/media-type-builder.js";
 import { BLANK_UUID } from "@/constants/constants.js";
