+++ mcp-server/src/umb-management-api/tools/media/__tests__/get-media-urls.test.ts
@@ -4,7 +4,7 @@ import { MediaTestHelper } from "./helpers/media-test-helper.js";
 import { jest } from "@jest/globals";
 import { BLANK_UUID } from "@/constants/constants.js";
 import { TemporaryFileBuilder } from "../../temporary-file/__tests__/helpers/temporary-file-builder.js";
-import { createSnapshotResult } from "@/helpers/test-utils.js";
+import { createSnapshotResult } from "@/test-helpers/create-snapshot-result.js";
 
 const TEST_MEDIA_NAME = "_Test Media URLs";
