+++ mcp-server/src/umb-management-api/tools/member-group/__tests__/get-member-group.test.ts
@@ -1,7 +1,7 @@
 import { getMemberGroupByIdParams } from "@/umb-management-api/umbracoManagementAPI.zod.js";
 import GetMemberGroupTool from "../get/get-member-group.js";
 import { MemberGroupBuilder } from "./helpers/member-group-builder.js";
-import { createSnapshotResult } from "@/helpers/test-utils.js";
+import { createSnapshotResult } from "@/test-helpers/create-snapshot-result.js";
 import { jest } from "@jest/globals";
 import { BLANK_UUID } from "@/constants/constants.js";
