+++ mcp-server/src/umb-management-api/tools/member-type/__tests__/get-member-type.test.ts
@@ -1,6 +1,6 @@
 import { MemberTypeTestHelper } from "./helpers/member-type-helper.js";
 import GetMemberTypeTool from "../get/get-member-type.js";
-import { createSnapshotResult } from "@/helpers/test-utils.js";
+import { createSnapshotResult } from "@/test-helpers/create-snapshot-result.js";
 import { jest } from "@jest/globals";
 import { MemberTypeBuilder } from "./helpers/member-type-builder.js";
 import { BLANK_UUID } from "@/constants/constants.js";
