+++ mcp-server/src/umb-management-api/tools/user-group/__tests__/get-user-group.test.ts
@@ -1,7 +1,7 @@
 import { getUserGroupByIdParams } from "@/umb-management-api/umbracoManagementAPI.zod.js";
 import GetUserGroupTool from "../get/get-user-group.js";
 import { UserGroupBuilder } from "./helpers/user-group-builder.js";
-import { createSnapshotResult } from "@/helpers/test-utils.js";
+import { createSnapshotResult } from "@/test-helpers/create-snapshot-result.js";
 import { jest } from "@jest/globals";
 import { BLANK_UUID } from "@/constants/constants.js";
