+++ mcp-server/src/umb-management-api/tools/webhook/__tests__/get-webhook-by-id.test.ts
@@ -1,6 +1,6 @@
 import { WebhookTestHelper } from "./helpers/webhook-helper.js";
 import GetWebhookByIdTool from "../get/get-webhook-by-id.js";
-import { createSnapshotResult } from "@/helpers/test-utils.js";
+import { createSnapshotResult } from "@/test-helpers/create-snapshot-result.js";
 import { jest } from "@jest/globals";
 import { WebhookBuilder } from "./helpers/webhook-builder.js";
 import { BLANK_UUID } from "@/constants/constants.js";
